[ppocr] simplify point2d

+++ projects/ppocr/mln/core/def/coord.hh
@@ -40,7 +40,7 @@ namespace mln
 
 
     /// Definition of the default coordinate type: 'short'.
-    typedef short coord;
+    typedef unsigned coord;
 
 
   } // end of namespace mln::def
